requirements.txt
torch==1.10.0
torchvision==0.11.0
numpy==1.18.0
opencv-python==4.5.0
scikit-image==0.18.0
tqdm==4.0.0
tensorboard==2.0.0
lpips==0.1.3
Pillow==8.0.0